Prepped repository for 2025-26 Battle Bots season
Moved old documentation into the `archived_24-25` directory to make space for the current ESW Battle Bots team.

Create `src` directory that contains basic code for moving the battle bot around. The intent is for the control matrix to contain values in the range [-1.0, 1.0] that will be used to better control the bot's response to the RC transmitter / receiver.

Added write-up from Hannah on the spring-force system inside this year's battle bot